requirements.txt
discord.py
aiosqlite
python-dotenv
psutil
chat-exporter
Pillow
matplotlib
quart
hypercorn